requirements.txt
django
requests
bs4
psycopg2
python-dotenv